backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfplumber==0.11.0
sqlalchemy==2.0.30
pydantic==2.10.6
python-multipart==0.0.9
openai==1.35.0
aiofiles==23.2.1
python-dotenv==1.0.1
supabase==2.5.0
PyJWT==2.8.0